Feature/restore cpc port (#67)
* [MY] restore the CPC version

Albeit working very slowly and kind of glitchy...but it's fully playable.

* [MY] improve framerate on CPC with direct memory access

* [MY] Further optimize graphical output on CPC

This also seems to fix some minor glitches.

* [MY] Fix text rendering on CPC

* [MY] refactor disk creation command to better divide responsabilities

+++ my_frontend/src/sdl2.c
@@ -62,7 +62,7 @@ void enterTextMode() {
 void exitTextMode() {
 }
 
-void vLine(uint8_t x0, uint8_t y0, uint8_t y1, uint shouldStipple) {
+void vLine(uint8_t x0, uint8_t y0, uint8_t y1, uint8_t shouldStipple) {
 	int16_t y;
 	int16_t _y0 = y0;
 	int16_t _y1 = y1;
